Extremely messy version of the typeahead working
Summary:
Adding arcconfig.

Typeahead works great, but I'm in the progress of cleaning up the code.

Typeahead for skills now works

Made changes to not use jquery

Fixed typeahead to not rely on jquery

New changes in typeahead to not use jquery

Test Plan:
I tested it briefly on localhost. Everything seemed good.

I tested it

Reviewers: DrAlexMV

Subscribers: nsundin

Differential Revision: http://beta.founderati.io/D4